pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oncology-agentic-rag"
version = "0.1.0"
description = "Agentic RAG scaffold for oncology"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "chromadb>=0.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]
